+++ pnlibrary-api/src/main/kotlin/ru/privatenull/pnlibrary/api/actions/Action.kt
@@ -11,6 +11,7 @@ import ru.privatenull.pnlibrary.api.config.ConfigTypes
     ConfigType(SoundAction::class, "sound"),
     ConfigType(ConsoleLogAction::class, "console", aliases = ["log"]),
     ConfigType(DelayAction::class, "delay", aliases = ["later"]),
+    ConfigType(ConditionalAction::class, "when", aliases = ["condition", "if"]),
     ConfigType(EffectAction::class, "effect", aliases = ["potion-effect"]),
     ConfigType(ParticleAction::class, "particle", aliases = ["particles"]),
 )