+++ pnlibrary-bungee/src/main/kotlin/ru/privatenull/pnlibrary/bungee/BungeeLibraryPlayer.kt
@@ -15,6 +15,7 @@ import java.util.UUID
 class BungeeLibraryPlayer private constructor(private val player: ProxiedPlayer) : ru.privatenull.pnlibrary.api.actions.LibraryPlayer {
     override val uniqueId: UUID get() = player.uniqueId
     override val name: String get() = player.name
+    override fun hasPermission(permission: String): Boolean = player.hasPermission(permission)
 
     override fun sendMessage(text: Component) {
         player.sendMessage(*TextComponent.fromLegacyText(LEGACY.serialize(text)))